Create Missing element
This Java program finds the missing element from a sequence of integers. It is assumed that the sequence should contain integers from 1 to N, but one integer is missing. The program calculates the sum of the numbers from 1 to N and subtracts the sum of the numbers provided in the array. The difference is the missing number.